Move indexForAt utility function
